Changed the wrap method and renamed it.
Changed the method that is supposed to wrap column or row indexes around
if the rover moves off grid. Also gave it the name wrap, as it's not
using modulus anymore. I had to fix this in the rotation logic too.

Also got rid of some comment code and started trimming individual inputs
when extracted from a character string.